requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
lighter-sdk==1.0.4
httpx==0.27.2
pydantic==2.9.2
